drop per Rector level configuration; use more clear config setup - easier to combine and include
