+++ easy-coding-standard.neon
@@ -30,7 +30,7 @@ checkers:
         extraAllowedClasses:
             - 'PhpParser\Node\*'
             - 'PhpParser\Comment\Doc'
-            - SplObjectStorage # temporary before adding to Symplify\CodingStandard package
+            - 'Rector\DeprecationExtractor\Deprecation\*'
 
 parameters:
     exclude_checkers:
