+++ src/Rector/Contrib/Nette/HtmlAddMethodRector.php
@@ -3,20 +3,9 @@
 namespace Rector\Rector\Contrib\Nette;
 
 use Rector\Rector\AbstractChangeMethodNameRector;
-use Rector\Rector\Set\SetNames;
 
 final class HtmlAddMethodRector extends AbstractChangeMethodNameRector
 {
-    public function getSetName(): string
-    {
-        return SetNames::NETTE;
-    }
-
-    public function sinceVersion(): float
-    {
-        return 2.4;
-    }
-
     /**
      * @return string[][]
      */
