+++ src/Rector/Contrib/PHPUnit/NamespaceClassRector.php
@@ -3,20 +3,12 @@
 namespace Rector\Rector\Contrib\PHPUnit;
 
 use Rector\Rector\AbstractClassReplacerRector;
-use Rector\Rector\Set\SetNames;
 
+/**
+ * Covers https://github.com/sebastianbergmann/phpunit/blob/master/ChangeLog-6.0.md#changed-1
+ */
 final class NamespaceClassRector extends AbstractClassReplacerRector
 {
-    public function getSetName(): string
-    {
-        return SetNames::PHPUNIT;
-    }
-
-    public function sinceVersion(): float
-    {
-        return 6.0;
-    }
-
     /**
      * @return string[]
      */
