+++ src/Rector/Contrib/PHP_CodeSniffer/NamespaceClassRector.php
@@ -3,27 +3,22 @@
 namespace Rector\Rector\Contrib\PHP_CodeSniffer;
 
 use Rector\Rector\AbstractClassReplacerRector;
-use Rector\Rector\Set\SetNames;
 
+/**
+ * Covers https://github.com/squizlabs/PHP_CodeSniffer/wiki/Version-3.0-Upgrade-Guide
+ */
 final class NamespaceClassRector extends AbstractClassReplacerRector
 {
-    public function getSetName(): string
-    {
-        return SetNames::PHP_CODE_SNIFER;
-    }
-
-    public function sinceVersion(): float
-    {
-        return 3.0;
-    }
-
     /**
      * @return string[]
      */
     protected function getOldToNewClasses(): array
     {
         return [
-            'PHP_CodeSniffer_Sniffs_Sniff' => 'PHP_CodeSniffer\Sniffs\Sniff',
+            'PHP_CodeSniffer_Sniffs_File' => 'PHP_CodeSniffer\Sniffs\Sniff',
+            'PHP_CodeSniffer_File' => 'PHP_CodeSniffer\Files\File',
+            'PHP_CodeSniffer_Tokens' => 'PHP_CodeSniffer\Util\Tokens',
+            'StandardName_Tests_Category_TestSniffUnitTest' => 'PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest',
         ];
     }
 }
