+++ src/Rector/Contrib/Symfony/CommandToConstructorInjectionRector.php
@@ -1,9 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Rector\Contrib\SymfonyExtra;
+namespace Rector\Rector\Contrib\Symfony;
 
+use Nette\Utils\Strings;
 use PhpParser\Node;
 use PhpParser\Node\Expr\MethodCall;
+use PhpParser\Node\Name\FullyQualified;
 use Rector\Builder\Class_\ClassPropertyCollector;
 use Rector\Builder\Kernel\ServiceFromKernelResolver;
 use Rector\Builder\Naming\NameResolver;
@@ -11,23 +13,30 @@ use Rector\Node\Attribute;
 use Rector\NodeAnalyzer\SymfonyContainerCallsAnalyzer;
 use Rector\NodeFactory\NodeFactory;
 use Rector\Rector\AbstractRector;
-use Rector\Rector\Set\SetNames;
 use Rector\Tests\Rector\Contrib\SymfonyExtra\GetterToPropertyRector\Source\LocalKernel;
 
 /**
- * Converts all:
- * $this->get('some_service') # where "some_service" is name of the service in container.
+ * Ref: https://github.com/symfony/symfony/blob/master/UPGRADE-4.0.md#console
  *
- * into:
- * $this->someService # where "someService" is type of the service
+ * Before:
+ * class MyCommand extends ContainerAwareCommand
+ *
+ * $this->getContainer()->get('some_service');
+ *
+ * After:
+ * class MyCommand extends Command
+ *
+ * public function construct(SomeService $someService)
+ * {
+ *     $this->someService = $someService;
+ * }
+ *
+ * ...
+ *
+ * $this->someService
  */
-final class GetterToPropertyRector extends AbstractRector
+final class CommandToConstructorInjectionRector extends AbstractRector
 {
-    /**
-     * @var NameResolver
-     */
-    private $nameResolver;
-
     /**
      * @var ServiceFromKernelResolver
      */
@@ -37,6 +46,11 @@ final class GetterToPropertyRector extends AbstractRector
      * @var ClassPropertyCollector
      */
     private $classPropertyCollector;
+
+    /**
+     * @var NameResolver
+     */
+    private $nameResolver;
 
     /**
      * @var NodeFactory
@@ -49,35 +63,43 @@ final class GetterToPropertyRector extends AbstractRector
     private $symfonyContainerCallsAnalyzer;
 
     public function __construct(
-        NameResolver $nameResolver,
         ServiceFromKernelResolver $serviceFromKernelResolver,
         ClassPropertyCollector $classPropertyCollector,
+        NameResolver $nameResolver,
         NodeFactory $nodeFactory,
         SymfonyContainerCallsAnalyzer $symfonyContainerCallsAnalyzer
     ) {
-        $this->nameResolver = $nameResolver;
         $this->serviceFromKernelResolver = $serviceFromKernelResolver;
         $this->classPropertyCollector = $classPropertyCollector;
+        $this->nameResolver = $nameResolver;
         $this->nodeFactory = $nodeFactory;
         $this->symfonyContainerCallsAnalyzer = $symfonyContainerCallsAnalyzer;
     }
 
     public function isCandidate(Node $node): bool
     {
+        $class = (string) $node->getAttribute(Attribute::CLASS_NAME);
+
+        if (! Strings::endsWith($class, 'Command')) {
+            return false;
+        }
+
         if (! $node instanceof MethodCall) {
             return false;
         }
 
-        return $this->symfonyContainerCallsAnalyzer->isThisCall($node);
+        return $this->symfonyContainerCallsAnalyzer->isGetContainerCall($node);
     }
 
     /**
-     * @param MethodCall $methodCallNode
+     * @param MethodCall $node
      */
-    public function refactor(Node $methodCallNode): ?Node
+    public function refactor(Node $node): ?Node
     {
+        $this->replaceParentContainerAwareCommandWithCommand($node);
+
         $serviceType = $this->serviceFromKernelResolver->resolveServiceClassFromArgument(
-            $methodCallNode->args[0],
+            $node->args[0],
             LocalKernel::class
         );
 
@@ -88,7 +110,7 @@ final class GetterToPropertyRector extends AbstractRector
         $propertyName = $this->nameResolver->resolvePropertyNameFromType($serviceType);
 
         $this->classPropertyCollector->addPropertyForClass(
-            (string) $methodCallNode->getAttribute(Attribute::CLASS_NAME),
+            (string) $node->getAttribute(Attribute::CLASS_NAME),
             $serviceType,
             $propertyName
         );
@@ -96,13 +118,9 @@ final class GetterToPropertyRector extends AbstractRector
         return $this->nodeFactory->createLocalPropertyFetch($propertyName);
     }
 
-    public function getSetName(): string
+    private function replaceParentContainerAwareCommandWithCommand(Node $node): void
     {
-        return SetNames::SYMFONY_EXTRA;
-    }
-
-    public function sinceVersion(): float
-    {
-        return 3.3;
+        $classNode = $node->getAttribute(Attribute::CLASS_NODE);
+        $classNode->extends = new FullyQualified('Symfony\Component\Console\Command\Command');
     }
 }
