+++ src/Rector/Contrib/Symfony/StringFormTypeToClassRector.php
@@ -3,11 +3,9 @@
 namespace Rector\Rector\Contrib\Symfony;
 
 use PhpParser\Node;
-use PhpParser\Node\Name;
 use PhpParser\Node\Scalar\String_;
 use Rector\NodeFactory\NodeFactory;
 use Rector\Rector\AbstractRector;
-use Rector\Rector\Set\SetNames;
 
 /**
  * Converts all:
@@ -63,16 +61,6 @@ final class StringFormTypeToClassRector extends AbstractRector
     public function __construct(NodeFactory $nodeFactory)
     {
         $this->nodeFactory = $nodeFactory;
-    }
-
-    public function getSetName(): string
-    {
-        return SetNames::SYMFONY;
-    }
-
-    public function sinceVersion(): float
-    {
-        return 4.0;
     }
 
     public function isCandidate(Node $node): bool
