+++ src/config/level/nette/nette21.yml
@@ -1,0 +1,3 @@
+checkers:
+    - Rector\Rector\Contrib\Nette\InjectPropertyRector
+    - Rector\Rector\Contrib\Nette\NetteConfiguratorRector
