+++ src/config/level/nette/nette23.yml
@@ -1,0 +1,2 @@
+checkers:
+    - Rector\Rector\Contrib\Nette\RemoveConfiguratorConstantsRector
