+++ src/config/level/php-code-sniffer/php-code-sniffer30.yml
@@ -1,0 +1,2 @@
+rectors:
+    - Rector\Rector\Contrib\PHP_CodeSniffer\NamespaceClassRector
