+++ src/config/level/phpunit/phpunit60.yml
@@ -1,0 +1,2 @@
+rectors:
+    - Rector\Rector\Contrib\PHPUnit\NamespaceClassRector
